Move logigng to util and some tests to subfolders
