+++ standalone/include/psen_scan_v2_standalone/data_conversion_layer/complete_scan_validator.h
@@ -19,7 +19,7 @@
 #include <boost/optional.hpp>
 
 #include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
-#include "psen_scan_v2_standalone/logging.h"
+#include "psen_scan_v2_standalone/util/logging.h"
 
 namespace psen_scan_v2_standalone
 {
