+++ standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
@@ -29,7 +29,7 @@
 #include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
 #include "psen_scan_v2_standalone/data_conversion_layer/raw_processing.h"
 #include "psen_scan_v2_standalone/raw_scanner_data.h"
-#include "psen_scan_v2_standalone/logging.h"
+#include "psen_scan_v2_standalone/util/logging.h"
 #include "psen_scan_v2_standalone/util/format_range.h"
 
 namespace psen_scan_v2_standalone
