+++ standalone/test/unit_tests/data_conversion_layer/unittest_monitoring_frame_msg.cpp
@@ -27,7 +27,7 @@
 #include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_deserialization.h"
 #include "psen_scan_v2_standalone/data_conversion_layer/raw_processing.h"
 
-#include "psen_scan_v2_standalone/istring_stream_builder.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/istring_stream_builder.h"
 #include "psen_scan_v2_standalone/communication_layer/udp_frame_dumps.h"
 #include "psen_scan_v2_standalone/data_conversion_layer/raw_data_array_conversion.h"
